[update] bugfix in intall.sh; generic update

+++ install.sh
@@ -1,7 +1,11 @@
 #!/bin/bash
+# who ran this script?
+U=${SUDO_USER:-$(whoami)}
 # where am i?
 base="$(dirname $(readlink -f $0))"
 optdir="/opt/oc.console/"
+
+. "$base/.utils/cmdexists.sh"
 
 # must run as super user
 if [ "$(id -u)" != "0" ]; then
@@ -19,11 +23,15 @@ if [ ! -d $optdir ]; then
    # create start icon
    cp "$optdir/oc.console.desktop" "/usr/share/applications/oc.console.desktop"
    # create desktop icon
-   cp "$optdir/oc.console.desktop" "/home/$USER/Desktop/oc.console.desktop"
+   cp "$optdir/oc.console.desktop" "/home/$U/Desktop/oc.console.desktop"
    echo "Installation completed"
 
    # install python requirements
-   python3 -m pip install prompt-toolkit
+   if [[ $(cmdexists "python3") -eq 1 ]]; then
+      python3 -m pip install prompt-toolkit
+   elif [[ $(cmdexists "python") -eq 1 ]]; then
+      python -m pip install prompt-toolkit
+   fi
 else
    echo "The program is already installed and located in /opt/oc.console"
 fi
